# Implement periodic cpu usage tracking in resource groups