--- a/presto-main/src/main/java/io/prestosql/execution/resourcegroups/InternalResourceGroupManager.java
+++ b/presto-main/src/main/java/io/prestosql/execution/resourcegroups/InternalResourceGroupManager.java
@@ -207,7 +207,7 @@ public final class InternalResourceGroupManager<C>
                 log.error(e, "Exception while generation cpu quota for %s", group);
             }
             try {
-                group.processQueuedQueries();
+                group.updateGroupsAndProcessQueuedQueries();
             }
             catch (RuntimeException e) {
                 log.error(e, "Exception while processing queued queries for %s", group);