--- a/presto-main/src/main/java/io/prestosql/server/ResourceGroupInfo.java
+++ b/presto-main/src/main/java/io/prestosql/server/ResourceGroupInfo.java
@@ -15,6 +15,7 @@ package io.prestosql.server;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.airlift.units.DataSize;
+import io.airlift.units.Duration;
 import io.prestosql.spi.resourcegroups.ResourceGroupId;
 import io.prestosql.spi.resourcegroups.ResourceGroupState;
 import io.prestosql.spi.resourcegroups.SchedulingPolicy;
@@ -43,6 +44,7 @@ public class ResourceGroupInfo
     private final int maxQueuedQueries;
 
     private final DataSize memoryUsage;
+    private final Duration cpuUsage;
     private final int numQueuedQueries;
     private final int numRunningQueries;
     private final int numEligibleSubGroups;
@@ -64,6 +66,7 @@ public class ResourceGroupInfo
             int maxQueuedQueries,
 
             DataSize memoryUsage,
+            Duration cpuUsage,
             int numQueuedQueries,
             int numRunningQueries,
             int numEligibleSubGroups,
@@ -85,6 +88,7 @@ public class ResourceGroupInfo
         this.maxQueuedQueries = maxQueuedQueries;
 
         this.memoryUsage = requireNonNull(memoryUsage, "memoryUsage is null");
+        this.cpuUsage = requireNonNull(cpuUsage, "cpuUsage is null");
         this.numQueuedQueries = numQueuedQueries;
         this.numRunningQueries = numRunningQueries;
         this.numEligibleSubGroups = numEligibleSubGroups;
@@ -128,6 +132,12 @@ public class ResourceGroupInfo
     public DataSize getMemoryUsage()
     {
         return memoryUsage;
+    }
+
+    @JsonProperty
+    public Duration getCpuUsage()
+    {
+        return cpuUsage;
     }
 
     @JsonProperty