--- a/presto-main/src/test/java/io/prestosql/execution/MockManagedQueryExecution.java
+++ b/presto-main/src/test/java/io/prestosql/execution/MockManagedQueryExecution.java
@@ -31,14 +31,15 @@ import java.util.List;
 import java.util.Optional;
 import java.util.OptionalDouble;
 
+import static com.google.common.base.Preconditions.checkState;
 import static io.airlift.units.DataSize.Unit.BYTE;
-import static io.airlift.units.DataSize.succinctBytes;
 import static io.prestosql.SystemSessionProperties.QUERY_PRIORITY;
 import static io.prestosql.execution.QueryState.FAILED;
 import static io.prestosql.execution.QueryState.FINISHED;
 import static io.prestosql.execution.QueryState.QUEUED;
 import static io.prestosql.execution.QueryState.RUNNING;
 import static io.prestosql.testing.TestingSession.testSessionBuilder;
+import static java.util.Objects.requireNonNull;
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
 import static java.util.concurrent.TimeUnit.NANOSECONDS;
 
@@ -46,33 +47,41 @@ public class MockManagedQueryExecution
         implements ManagedQueryExecution
 {
     private final List<StateChangeListener<QueryState>> listeners = new ArrayList<>();
-    private final DataSize memoryUsage;
-    private final Duration cpuUsage;
     private final Session session;
+
+    private DataSize memoryUsage;
+    private Duration cpuUsage;
     private QueryState state = QUEUED;
     private Throwable failureCause;
 
-    public MockManagedQueryExecution(long memoryUsage)
+    private MockManagedQueryExecution(String queryId, int priority, DataSize memoryUsage, Duration cpuUsage)
     {
-        this(memoryUsage, "query_id", 1);
-    }
-
-    public MockManagedQueryExecution(long memoryUsage, String queryId, int priority)
-    {
-        this(memoryUsage, queryId, priority, new Duration(0, MILLISECONDS));
-    }
-
-    public MockManagedQueryExecution(long memoryUsage, String queryId, int priority, Duration cpuUsage)
-    {
-        this.memoryUsage = succinctBytes(memoryUsage);
-        this.cpuUsage = cpuUsage;
+        requireNonNull(queryId, "queryId is null");
         this.session = testSessionBuilder()
+                .setQueryId(QueryId.valueOf(queryId))
                 .setSystemProperty(QUERY_PRIORITY, String.valueOf(priority))
                 .build();
+
+        this.memoryUsage = requireNonNull(memoryUsage, "memoryUsage is null");
+        this.cpuUsage = requireNonNull(cpuUsage, "cpuUsage is null");
+    }
+
+    public void consumeCpuTime(Duration cpuTimeDelta)
+    {
+        checkState(state == RUNNING, "cannot consume CPU in a non-running state");
+        long newCpuTime = cpuUsage.toMillis() + cpuTimeDelta.toMillis();
+        this.cpuUsage = new Duration(newCpuTime, MILLISECONDS);
+    }
+
+    public void setMemoryUsage(DataSize memoryUsage)
+    {
+        checkState(state == RUNNING, "cannot set memory usage in a non-running state");
+        this.memoryUsage = memoryUsage;
     }
 
     public void complete()
     {
+        memoryUsage = new DataSize(0, BYTE);
         state = FINISHED;
         fireStateChange();
     }
@@ -120,11 +129,11 @@ public class MockManagedQueryExecution
                         new DataSize(14, BYTE),
                         15,
                         16.0,
-                        new DataSize(17, BYTE),
-                        new DataSize(18, BYTE),
+                        memoryUsage,
+                        memoryUsage,
                         new DataSize(19, BYTE),
                         new DataSize(20, BYTE),
-                        new Duration(21, NANOSECONDS),
+                        cpuUsage,
                         new Duration(22, NANOSECONDS),
                         false,
                         ImmutableSet.of(),
@@ -166,6 +175,7 @@ public class MockManagedQueryExecution
     @Override
     public void fail(Throwable cause)
     {
+        memoryUsage = new DataSize(0, BYTE);
         state = FAILED;
         failureCause = cause;
         fireStateChange();
@@ -187,6 +197,45 @@ public class MockManagedQueryExecution
     {
         for (StateChangeListener<QueryState> listener : listeners) {
             listener.stateChanged(state);
+        }
+    }
+
+    public static class MockManagedQueryExecutionBuilder
+    {
+        private DataSize memoryUsage = new DataSize(0, BYTE);
+        private Duration cpuUsage = new Duration(0, MILLISECONDS);
+        private int priority = 1;
+        private String queryId = "query_id";
+
+        public MockManagedQueryExecutionBuilder() {}
+
+        public MockManagedQueryExecutionBuilder withInitialMemoryUsage(DataSize memoryUsage)
+        {
+            this.memoryUsage = memoryUsage;
+            return this;
+        }
+
+        public MockManagedQueryExecutionBuilder withInitialCpuUsage(Duration cpuUsage)
+        {
+            this.cpuUsage = cpuUsage;
+            return this;
+        }
+
+        public MockManagedQueryExecutionBuilder withPriority(int priority)
+        {
+            this.priority = priority;
+            return this;
+        }
+
+        public MockManagedQueryExecutionBuilder withQueryId(String queryId)
+        {
+            this.queryId = queryId;
+            return this;
+        }
+
+        public MockManagedQueryExecution build()
+        {
+            return new MockManagedQueryExecution(queryId, priority, memoryUsage, cpuUsage);
         }
     }
 }