This PR contains the following changes at a high level:

* Periodically update and cache cpu usage for resource groups.
* Add memory and cpu usage update at query completion.
* Update eligibility through a post-order traversal in `updateEligibilityAll` after resource usage update to avoid calling `updateEligibility` recursively in the `internalRefreshStats`.
* Rename `processQueuedQueries` to reflect what the method is doing
* move memory limit check in `canRunMore` for better readability.

Resolves #1059 . @dain @wagnermarkd