@phd3 2019-08-03
@dain @raghavsethi I've updated the PR addressing your comments.

Numbers for TestResourceGroups benchmark:

Benchmark | (children) | (queries) | Time taken (us/op) (Before) | Time taken (us/op) (With `updateEligibilityInSubTree`method) | Time taken (us/op)(After removing `updateEligibilityInSubTree` method as in `c16e711 `)
-- | -- | -- | -- | -- | --
BenchmarkResourceGroup.benchmark | 1000 | 100 | 1.364 | 216.476 | 2.499
BenchmarkResourceGroup.benchmark | 1000 | 1000 | 13.428 | 236.021 | 33.543
BenchmarkResourceGroup.benchmark | 1000 | 10000 | 495.993 | 1148.303 | 1211.227
BenchmarkResourceGroup.benchmark | 10000 | 100 | 1.094 | 2353.871 | 2.405
BenchmarkResourceGroup.benchmark | 10000 | 1000 | 12.18 | 2489.451 | 26.708
BenchmarkResourceGroup.benchmark | 10000 | 10000 | 445.518 | 4522.486 | 717.43
BenchmarkResourceGroup.benchmark | 100000 | 100 | 1.217 | 33370.087 | 2.151
BenchmarkResourceGroup.benchmark | 100000 | 1000 | 14.851 | 31872.003 | 26.752
BenchmarkResourceGroup.benchmark | 100000 | 10000 | 455.631 | 32303.722 | 534.325

After seeing that the benchmark takes a lot longer to execute with `updateEligibilityInSubTree` method, I replaced its usage with `updateEligibilityOnPathToRoot` calls in `internalCpuQuotaGeneration` and `updateResourceUsageAndGetDelta`.  This change (a bit more than) doubles the time taken for the update (from before this PR), which seems reasonable to me given that we're updating two values per node now.

Removing `updateEligibilityInSubTree` requires us to invoke `updateEligibilityOnPathToRoot` method inside and `internalGenerateCpuQuota`. This works well for shallow trees, even with a high fanout. But I've a concern that it might not work well for deep trees. Do you think we should optimize this further?