@phd3 2019-08-05
@raghavsethi I think `updateEligibilityInSubtree` method created some confusion. I've removed it [(here](https://github.com/prestosql/presto/pull/1128/commits/c16e711f5f76fc1e81f562804c6705579658468e) looking at the benchmark result. That might simplify things, and work well for shallow trees at least.  Happy to schedule a call if required for discussing this further.