@raghavsethi 2019-08-05
If that's the case, it seems like we one have one `updateEligibility` method, can we rename it back?