@raghavsethi 2019-08-08
Minor changes requested, looks generally good.