@phd3 2019-08-09
@raghavsethi Thanks for the review, I've updated the PR addressing your latest comments.

> It's definitely not great behavior and could lead to a lot of user confusion, did your change make this worse? If not, then you can address it in a follow up PR.

This has been the behavior before the change as well. But it starves a queued query only if new queries submitted between the invocation of cpu quota generation and updateGroupsAndProcessQueuedQueries exceed limits instantaneously.