@raghavsethi 2019-11-23
Could you provide your resource groups config file?